array intervall: move rttm stuff to a new file
